// src/app/models/crud/proveedor.ts
export class Proveedor {
  id?: number = 0;
  nit: string | null | undefined = "";
  nombre: string | null | undefined = "";
  encargado: string | null | undefined = "";
  telefono: string | null | undefined = "";
  correo: string | null | undefined = "";
  direccion: string | null | undefined = "";
  departamento: string | null | undefined = "";
  municipio: string | null | undefined = "";
  rubros: string | null | undefined = "";
  estado: string | null | undefined = "";
}